requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
rich>=13.7.0
